De-polymorphism graph persistence API
