+++ babel.config.json
@@ -1,4 +1,4 @@
 {
   "presets": ["@babel/preset-env", ["@babel/preset-react", { "runtime": "automatic" }], "@babel/preset-typescript"],
-  "plugins": [["@babel/plugin-proposal-decorators", { "legacy": true }]]
+  "plugins": [["@babel/plugin-proposal-decorators", { "legacy": true }], "babel-plugin-parameter-decorator"]
 }
