+++ package.json
@@ -36,6 +36,8 @@
   "dependencies": {
     "@logto/react": "^2.1.1",
     "dotenv": "^16.3.1",
+    "inversify": "^6.0.1",
+    "inversify-inject-decorators": "^3.1.0",
     "lodash-es": "^4.17.21",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.0",
@@ -44,12 +46,13 @@
     "react-router-dom": "^6.8.1",
     "react-spring": "^9.7.3",
     "redux": "^3.7.2",
+    "reflect-metadata": "^0.1.13",
     "styled-components": "^5.3.9"
   },
   "devDependencies": {
     "@babel/cli": "^7.21.0",
     "@babel/core": "^7.21.0",
-    "@babel/plugin-proposal-decorators": "^7.22.10",
+    "@babel/plugin-proposal-decorators": "^7.23.2",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.21.0",
@@ -62,6 +65,7 @@
     "@typescript-eslint/parser": "^5.61.0",
     "babel-jest": "^27.4.2",
     "babel-loader": "^9.1.2",
+    "babel-plugin-parameter-decorator": "^1.0.16",
     "css-loader": "^6.7.3",
     "cypress": "^12.17.3",
     "eslint": "^8.44.0",
