+++ packages/nexusgraph-app/inversify.config.ts
@@ -1,8 +1,6 @@
 // Copyright 2023 Paion Data. All rights reserved.
 import { Container } from "inversify";
 
-import { AstraiosClient } from "../nexusgraph-astraios";
-import { GraphQlClient } from "../nexusgraph-astraios/src/GraphQlClient";
 import { NLPClient, TheresaClient } from "../nexusgraph-nlp";
 
 /**
@@ -25,6 +23,5 @@ const container = new Container();
  * Bind the class we use to implement the interface
  */
 container.bind<NLPClient>(TYPES.NLPClient).to(TheresaClient).inSingletonScope();
-container.bind<AstraiosClient>(TYPES.AstraiosClient).to(GraphQlClient).inSingletonScope();
 
 export { container, TYPES };
