+++ packages/nexusgraph-app/src/DevApp.tsx
@@ -1,7 +1,6 @@
 // Copyright 2023 Paion Data. All rights reserved.
 import { useDispatch } from "react-redux";
 import { updateOAuthState } from "../../nexusgraph-redux";
-import { container, TYPES } from "../inversify.config";
 import App from "./App";
 
 export default function DevApp(): JSX.Element {
@@ -17,8 +16,6 @@ export default function DevApp(): JSX.Element {
       userInfo: devUserInfo,
     })
   );
-  container.bind<string>(TYPES.accessToken).toConstantValue(devToken);
-  container.bind<string>(TYPES.userId).toConstantValue(devUserId);
 
   return <App />;
 }
