+++ packages/nexusgraph-astraios/index.ts
@@ -1,4 +1,3 @@
 // Copyright 2023 Paion Data. All rights reserved.
 
-export type { AstraiosClient } from "./src/AstraiosClient";
-export { JsonApiAstraiosClient } from "./src/JsonApiClient";
+export { AstraiosClient } from "./src/AstraiosClient";
