+++ packages/nexusgraph-nlp/package.json
@@ -1,10 +1,5 @@
 {
   "name": "@paiondata/nexusgraph-nlp",
   "version": "0.1.0",
-  "description": "Natural language processing",
-  "dependencies": {
-    "inversify": "^6.0.1",
-    "inversify-inject-decorators": "^3.1.0",
-    "reflect-metadata": "^0.1.13"
-  }
+  "description": "Natural language processing"
 }
